Remove the doc infos - they are now integrated with the persistent ones. We'll get nn-persistent ones back for the next checkin, but only as placeholds to remember that a type of that name exists. Refactor some of THtml's long functions. Simplify TClassDoc's constructor.
git-svn-id: https://root.cern.ch/svn/root/branches/dev@30901 27541ba8-7e3a-0410-8455-c3a389f83636

+++ axel/html/src/TDocumented.cxx
@@ -51,7 +51,7 @@ Int_t Doc::TDocumented::Compare(const TObject* obj) const {
    };
    TClass* me = IsA();
    TClass* it = obj->IsA();
-   for (int i = 0; i < sizeof(sClasses) / sizeof(TClass*); ++i) {
+   for (size_t i = 0; i < sizeof(sClasses) / sizeof(TClass*); ++i) {
       if (me == sClasses[i]) {
          return -1;
       }
